Fix python client and remove sqlmodel from dependencies
