+++ .pre-commit-config.yaml
@@ -110,7 +110,7 @@ repos:
       entry: pdm lock --check --project .
       files: ^pyproject.toml$
 - repo: https://github.com/pdm-project/pdm
-  rev: 2.8.1
+  rev: 2.11.1
   hooks:
     - id: pdm-lock-check
       name: pdm-lock-check-server
