+++ datajunction-server/datajunction_server/api/attributes.py
@@ -134,7 +134,7 @@ def default_attribute_types(session: Session = Depends(get_session)):
     )
     attribute_types = session.execute(statement).scalars().all()
     for type_ in attribute_types:
-        if type_:
+        if type_:  # pragma: no cover
             type_.name = default_attribute_type_names[type_.name].name
             type_.namespace = default_attribute_type_names[type_.name].namespace
             type_.description = default_attribute_type_names[type_.name].description
